Second

+++ app/Http/Controllers/PostsController.php
@@ -2,11 +2,22 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Post;
+use Illuminate\Http\Request;
+use DB;
 
 class PostsController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +25,8 @@ class PostsController extends Controller
      */
     public function index()
     {
-      $posts = Post::all();
+      //$posts = Post::all();
+      $posts = DB::select('SELECT *FROM posts');
       return view('posts.index')->with('posts', $posts);
     }
 
@@ -44,7 +56,7 @@ class PostsController extends Controller
       $post = new Post;
       $post->title = $request->input('title');
       $post->body = $request->input('body');
-      //$post->user_id = auth()->user()->id;
+      $post->user_id = auth()->user()->id;
       $post->save();
 
       return redirect('/posts')->with('success', 'Статтю створено.');
@@ -71,6 +83,9 @@ class PostsController extends Controller
     public function edit($id)
     {
       $post = Post::find($id);
+      if(auth()->user()->id !== $post->user_id){
+        return redirect('/posts')->with('error', 'Немає доступу до цієї сторінки!');
+      }
       return view('posts.edit')->with('post', $post);
     }
 
@@ -105,6 +120,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
       $post = Post::find($id);
+
+      if(auth()->user()->id !== $post->user_id){
+        return redirect('/posts')->with('error', 'Немає доступу до цієї сторінки!');
+      }
+
       $post->delete();
       return redirect('/posts')->with('success', 'Статтю видалено.');
     }
